+++ src/pararnn/cells/__init__.py
@@ -1,5 +1,6 @@
 from pararnn.cells.para_cfc import ParaCfC
 from pararnn.cells.para_gru import ParaGRU
+from pararnn.cells.para_hopfield import ParaHopfield
 from pararnn.cells.para_lstm import ParaLSTM
 from pararnn.cells.para_m2rnn import ParaM2RNN
 from pararnn.cells.para_nlru import ParaNLRU
@@ -9,6 +10,7 @@ from pararnn.cells.protocol import RNNCell, check_cell
 __all__ = [
     "ParaCfC",
     "ParaGRU",
+    "ParaHopfield",
     "ParaLSTM",
     "ParaM2RNN",
     "ParaNLRU",